+++ src/main/java/com/openggf/game/sonic1/objects/Sonic1MzBrickObjectInstance.java
@@ -8,12 +8,11 @@ import com.openggf.graphics.RenderPriority;
 import com.openggf.level.objects.AbstractObjectInstance;
 import com.openggf.level.objects.ObjectArtKeys;
 import com.openggf.level.objects.ObjectSpawn;
-import com.openggf.level.objects.RewindRecreateContext;
-import com.openggf.level.objects.RewindRecreatable;
 import com.openggf.level.objects.SolidContact;
 import com.openggf.level.objects.SolidObjectListener;
 import com.openggf.level.objects.SolidObjectParams;
 import com.openggf.level.objects.SolidObjectProvider;
+import com.openggf.level.objects.SpawnRewindRecreatable;
 import com.openggf.level.objects.SubpixelMotion;
 import com.openggf.level.render.PatternSpriteRenderer;
 import com.openggf.physics.ObjectTerrainUtils;
@@ -43,7 +42,7 @@ import java.util.List;
  * Reference: docs/s1disasm/_incObj/46 MZ Bricks.asm
  */
 public class Sonic1MzBrickObjectInstance extends AbstractObjectInstance
-        implements SolidObjectProvider, SolidObjectListener, RewindRecreatable {
+        implements SolidObjectProvider, SolidObjectListener, SpawnRewindRecreatable {
 
     // From disassembly: move.w #$1B,d1
     private static final int HALF_WIDTH = 0x1B;
@@ -110,11 +109,6 @@ public class Sonic1MzBrickObjectInstance extends AbstractObjectInstance
         this.yVelocity = 0;
 
         updateDynamicSpawn(x, y);
-    }
-
-    @Override
-    public Sonic1MzBrickObjectInstance recreateForRewind(RewindRecreateContext ctx) {
-        return new Sonic1MzBrickObjectInstance(ctx.spawn());
     }
 
     @Override