+++ src/main/java/com/openggf/game/sonic1/objects/Sonic1PoleThatBreaksObjectInstance.java
@@ -11,8 +11,7 @@ import com.openggf.graphics.RenderPriority;
 import com.openggf.level.objects.AbstractObjectInstance;
 import com.openggf.level.objects.ObjectArtKeys;
 import com.openggf.level.objects.ObjectSpawn;
-import com.openggf.level.objects.RewindRecreateContext;
-import com.openggf.level.objects.RewindRecreatable;
+import com.openggf.level.objects.SpawnRewindRecreatable;
 import com.openggf.level.objects.TouchActorContextPolicy;
 import com.openggf.level.objects.TouchAttackBouncePolicy;
 import com.openggf.level.objects.TouchCategoryDecodeMode;
@@ -36,7 +35,7 @@ import java.util.List;
  * Disassembly reference: docs/s1disasm/_incObj/0B Pole that Breaks.asm
  */
 public class Sonic1PoleThatBreaksObjectInstance extends AbstractObjectInstance
-        implements TouchResponseProvider, TouchResponseListener, RewindRecreatable {
+        implements TouchResponseProvider, TouchResponseListener, SpawnRewindRecreatable {
 
     // move.w #make_art_tile(ArtTile_LZ_Pole,2,0),obGfx(a0)
     private static final int DISPLAY_PRIORITY = 4;
@@ -88,11 +87,6 @@ public class Sonic1PoleThatBreaksObjectInstance extends AbstractObjectInstance
         super(spawn, "PoleThatBreaks");
         int subtype = spawn.subtype() & 0xFF;
         this.poleTime = subtype * 60;
-    }
-
-    @Override
-    public Sonic1PoleThatBreaksObjectInstance recreateForRewind(RewindRecreateContext ctx) {
-        return new Sonic1PoleThatBreaksObjectInstance(ctx.spawn());
     }
 
     @Override