+++ src/main/java/com/openggf/game/sonic1/objects/Sonic1SceneryObjectInstance.java
@@ -4,8 +4,7 @@ import com.openggf.graphics.GLCommand;
 import com.openggf.level.objects.AbstractObjectInstance;
 import com.openggf.level.objects.ObjectArtKeys;
 import com.openggf.level.objects.ObjectSpawn;
-import com.openggf.level.objects.RewindRecreateContext;
-import com.openggf.level.objects.RewindRecreatable;
+import com.openggf.level.objects.SpawnRewindRecreatable;
 import com.openggf.level.render.PatternSpriteRenderer;
 
 import java.util.List;
@@ -26,7 +25,7 @@ import java.util.List;
  * <p>
  * Reference: docs/s1disasm/_incObj/1C Scenery.asm
  */
-public class Sonic1SceneryObjectInstance extends AbstractObjectInstance implements RewindRecreatable {
+public class Sonic1SceneryObjectInstance extends AbstractObjectInstance implements SpawnRewindRecreatable {
 
     // From Scen_Values: subtype 3 uses Map_Bri frame 1 (bridge stump)
     private static final int SUBTYPE_BRIDGE_STUMP = 3;
@@ -50,11 +49,6 @@ public class Sonic1SceneryObjectInstance extends AbstractObjectInstance implemen
             artKey = ObjectArtKeys.SCENERY;
             frameIndex = 0;
         }
-    }
-
-    @Override
-    public Sonic1SceneryObjectInstance recreateForRewind(RewindRecreateContext ctx) {
-        return new Sonic1SceneryObjectInstance(ctx.spawn());
     }
 
     @Override