+++ src/main/java/com/openggf/game/sonic1/objects/Sonic1VanishingPlatformObjectInstance.java
@@ -8,13 +8,12 @@ import com.openggf.level.LevelManager;
 import com.openggf.level.objects.AbstractObjectInstance;
 import com.openggf.level.objects.ObjectArtKeys;
 import com.openggf.level.objects.ObjectSpawn;
-import com.openggf.level.objects.RewindRecreateContext;
-import com.openggf.level.objects.RewindRecreatable;
 import com.openggf.level.objects.SolidContact;
 import com.openggf.level.objects.SolidExecutionMode;
 import com.openggf.level.objects.SolidObjectListener;
 import com.openggf.level.objects.SolidObjectParams;
 import com.openggf.level.objects.SolidObjectProvider;
+import com.openggf.level.objects.SpawnRewindRecreatable;
 import com.openggf.level.render.PatternSpriteRenderer;
 import com.openggf.sprites.playable.AbstractPlayableSprite;
 import com.openggf.game.PlayableEntity;
@@ -50,7 +49,7 @@ import java.util.List;
  * Reference: docs/s1disasm/_incObj/6C SBZ Vanishing Platforms.asm
  */
 public class Sonic1VanishingPlatformObjectInstance extends AbstractObjectInstance
-        implements SolidObjectProvider, SolidObjectListener, RewindRecreatable {
+        implements SolidObjectProvider, SolidObjectListener, SpawnRewindRecreatable {
 
     // From disassembly: move.b #$10,obActWid(a0)
     private static final int HALF_WIDTH = 0x10;
@@ -142,11 +141,6 @@ public class Sonic1VanishingPlatformObjectInstance extends AbstractObjectInstanc
         this.animFrameTimer = ANIM_FRAME_DURATION;
         this.currentFrame = 0;
         this.playerStanding = false;
-    }
-
-    @Override
-    public Sonic1VanishingPlatformObjectInstance recreateForRewind(RewindRecreateContext ctx) {
-        return new Sonic1VanishingPlatformObjectInstance(ctx.spawn());
     }
 
     @Override